profile-module2 fix mr comments
