+++ FakeNFT/Scenes/Common/Views/RatingView.swift
@@ -30,9 +30,7 @@ final class RatingView: UIView {
         setupUI()
     }
 
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
+    required init?(coder: NSCoder) { nil }
 
     // MARK: - Setup UI
