+++ FakeNFT/Scenes/NFTCollectionDetail/Cell/NFTCollectionViewCell.swift
@@ -79,9 +79,8 @@ final class NFTCollectionViewCell: UICollectionViewCell, ReuseIdentifying {
         setupLayout()
     }
 
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
+    required init?(coder: NSCoder) { nil
+}
 
     override func prepareForReuse() {
         super.prepareForReuse()
