+++ FakeNFT/Scenes/Profile/ProfileInteractor.swift
@@ -2,6 +2,7 @@ import Foundation
 
 protocol ProfileInteractorProtocol: AnyObject {
     func fetchProfile(userId: String)
+    func saveProfile(_ profile: Profile)
 }
 
 final class ProfileInteractor: ProfileInteractorProtocol {
@@ -21,6 +22,18 @@ final class ProfileInteractor: ProfileInteractorProtocol {
                 self.presenter?.didFetchProfile(profile)
             case .failure(let error):
                 self.presenter?.didFailFetchingProfile(error: error)
+            }
+        }
+    }
+    
+    func saveProfile(_ profile: Profile) {
+        profileService.updateProfile(profile) { [weak self] result in
+            guard let self = self else { return }
+            switch (result) {
+            case .success(let profile) :
+                self.presenter?.didFetchProfile(profile)
+            case .failure(let error):
+                self.presenter?.didFailUpdateProfile(error: error)
             }
         }
     }
