+++ FakeNFT/Scenes/Profile/ProfilePresenter.swift
@@ -12,11 +12,11 @@ final class LoadingState: ProfileState {
 
 final class LoadedState: ProfileState {
     private let profile: Profile
-
+    
     init(profile: Profile) {
         self.profile = profile
     }
-
+    
     func handle(presenter: ProfilePresenter) {
         presenter.view?.hideLoading()
         presenter.view?.showProfile(profile)
@@ -25,11 +25,11 @@ final class LoadedState: ProfileState {
 
 final class ErrorState: ProfileState {
     private let error: Error
-
+    
     init(error: Error) {
         self.error = error
     }
-
+    
     func handle(presenter: ProfilePresenter) {
         presenter.view?.hideLoading()
         presenter.view?.showError(error.localizedDescription)
@@ -43,6 +43,7 @@ protocol ProfilePresenterProtocol: AnyObject {
     func didFailFetchingProfile(error: Error)
     func didTapEditButton()
     func didTapMyNftButton()
+    func didFailUpdateProfile(error: Error)
 }
 
 final class ProfilePresenter {
@@ -52,20 +53,11 @@ final class ProfilePresenter {
     
     private let userId: String
     private var state: ProfileState = LoadingState()
-
+    
     init(userId: String, interactor: ProfileInteractorProtocol, router: ProfileRouterProtocol) {
         self.userId = userId
         self.interactor = interactor
         self.router = router
-    }
-    
-    deinit {
-        NotificationCenter.default.removeObserver(self, name: .profileDidChange, object: nil)
-    }
-    
-    @objc private func handleProfileDidChange(_ notification: Notification) {
-        guard let updatedProfile = notification.object as? Profile else { return }
-        view?.showProfile(updatedProfile)
     }
 }
 
@@ -73,26 +65,32 @@ extension ProfilePresenter: ProfilePresenterProtocol {
     func viewDidLoad() {
         state.handle(presenter: self)
         interactor.fetchProfile(userId: userId)
-        NotificationCenter.default.addObserver(
-                    self,
-                    selector: #selector(handleProfileDidChange(_:)),
-                    name: .profileDidChange,
-                    object: nil
-                )
     }
     
     func didFetchProfile(_ profile: Profile) {
         state = LoadedState(profile: profile)
         state.handle(presenter: self)
     }
-
+    
     func didFailFetchingProfile(error: Error) {
         state = ErrorState(error: error)
         state.handle(presenter: self)
     }
     
     func didTapEditButton() {
-        router.openProfileEdit(withUserId: userId)
+        router.openProfileEdit(withUserId: userId) { [weak self] profile in
+            guard let self else {
+                return
+            }
+            self.didFetchProfile(profile)
+            self.interactor.saveProfile(profile)
+        }
+    }
+    
+    func didFailUpdateProfile(error: any Error) {
+        state = ErrorState(error: error)
+        state.handle(presenter: self)
+        interactor.fetchProfile(userId: userId)
     }
     
     func didTapMyNftButton() {
