+++ FakeNFT/Scenes/ProfileEdit/Model/ProfileSaveData.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+struct ProfileSaveData {
+    var name: String
+    var description: String
+    var website: String
+    var photo: UIImage?
+}
