+++ FakeNFT/Scenes/ProfileEdit/ProfileEditInteractorInput.swift
@@ -1,6 +1,5 @@
 protocol ProfileEditInteractorInput: AnyObject {
     func loadProfile()
-    func saveProfile(_ profile: Profile)
 }
 
 final class ProfileEditInteractor: ProfileEditInteractorInput {
@@ -20,18 +19,6 @@ final class ProfileEditInteractor: ProfileEditInteractorInput {
                 self.presenter?.profileLoaded(profile)
             case .failure(let error):
                 self.presenter?.profileLoadFailed(error: error)
-            }
-        }
-    }
-
-    func saveProfile(_ profile: Profile) {
-        profileService.updateProfile(profile) { [weak self] result in
-            guard let self = self else { return }
-            switch (result) {
-            case .success(_) :
-                self.presenter?.profileSaveSucceeded()
-            case .failure(let error):
-                self.presenter?.profileSaveFailed(error: error)
             }
         }
     }
