+++ FakeNFT/Services/Requests/UpdateProfileDto.swift
@@ -5,11 +5,8 @@ struct UpdateProfileDto: Dto {
 
     func asDictionary() -> [String: String] {
         return ["name": profile.name,
-                "avatar": profile.avatar,
                 "description": profile.description,
                 "website": profile.website,
-                "nfts": profile.nfts.joined(separator: ","),
-                "likes": profile.likes.joined(separator: ","),
-                "id": profile.id]
+                "likes": profile.likes.joined(separator: ",")]
     }
 }
